# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.GetBalanceOUT:
    properties:
      balance:
        type: string
      base:
        type: string
      ok:
        type: boolean
    type: object
...
  httputils.Error:
    properties:
      err:
        type: string
      ok:
        type: boolean
    type: object
host: 0.0.0.0:8080
info:
  contact:
    email: illiadimura@gmail.com
    name: Developer
    url: https://github.com/illiafox
  description: Balance API.
  license:
    name: Boost Software License 1.0
    url: https://opensource.org/licenses/BSL-1.0
  termsOfService: https://swagger.io/terms/
  title: Balance API
  version: "1.0"
paths:
...
  /user/{id}:
    get:
      description: Get balance by User ID
      parameters:
      - description: user id
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: currency abbreviation
        in: query
        minLength: 3
        name: base
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Balance data
          schema:
            allOf:
            - $ref: '#/definitions/dto.GetBalanceOUT'
            - properties:
                balance:
                  type: integer
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputils.Error'
        "406":
          description: Not Acceptable
          schema:
            $ref: '#/definitions/httputils.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.Error'
      summary: Get user balance
      tags:
      - balance
...
schemes:
- http
- https
swagger: "2.0"